Move MSpec formatters to a separate file
